requirements.txt
Flask 
Flask-SQLAlchemy 
psycopg2-binary 
python-dotenv 
Flask-Cors
Flask-Migrate>=4.0.7

